revert: format

+++ tracer/src/instruction/ld.rs
@@ -16,7 +16,8 @@ declare_riscv_instr!(
 impl LD {
     fn exec(&self, cpu: &mut Cpu, ram_access: &mut <LD as RISCVInstruction>::RAMAccess) {
         // The LD instruction loads a 64-bit value from memory into register rd for RV64I.
-        let address = (cpu.x[self.operands.rs1] as u64).wrapping_add(self.operands.imm);
+        let address =
+            (cpu.x[self.operands.rs1] as u64).wrapping_add(self.operands.imm as i32 as u64);
         let value = cpu.get_mut_mmu().load_doubleword(address);
         cpu.x[self.operands.rd] = match value {
             Ok((value, memory_read)) => {
